+++ app/src/main/java/app/android/damien/reef/model/apex/Media.kt
@@ -2,7 +2,7 @@ package app.android.damien.reef.model.apex
 
 data class Media(
     val _id: String,
-    val auth: Int,
+    val auth: Long,
     val index: Int,
     val name: String,
     val type: String,
